Actualizacion hasta 16/11
