+++ src/main/java/ProyectoFinal04/Empender/Servicios/UsuarioServicio.java
@@ -64,6 +64,7 @@ public class UsuarioServicio implements UserDetailsService {
         
         //
         return repositorioUsuario.save(user);
+        
     }
 
     @Transactional
